+++ core/designsystem/build.gradle.kts
@@ -31,6 +31,7 @@ kotlin {
             implementation(compose.components.resources)
             implementation(compose.material3)
             implementation(compose.ui)
+            implementation(libs.coil.compose)
             implementation(libs.compose.material3.windowsizeclass)
             implementation(libs.material.kolor)
         }
