+++ core/designsystem/src/commonMain/kotlin/com/adammcneilly/spacenerd/core/designsystem/models/ImageModel.kt
@@ -1,23 +1,23 @@
-package com.adammcneilly.spacenerd.core.displaymodels
+package com.adammcneilly.spacenerd.core.designsystem.models
 
 import org.jetbrains.compose.resources.DrawableResource
 
-sealed interface ImageDisplayModel {
+sealed interface ImageModel {
     data class Remote(
         val url: String,
-    ) : ImageDisplayModel
+    ) : ImageModel
 
     data class AndroidLocal(
         val drawableRes: Int,
-    ) : ImageDisplayModel
+    ) : ImageModel
 
     data class Local(
         val resource: DrawableResource,
-    ) : ImageDisplayModel
+    ) : ImageModel
 
     /**
      * Typically only used in previews, this will be used
      * when we don't have an actual image to supply.
      */
-    data object Placeholder : ImageDisplayModel
+    data object Placeholder : ImageModel
 }
